Updated sub_processes to record parts of minutes as Floats instead of Integer values

+++ app/models/sub_process.rb
@@ -4,8 +4,8 @@ class SubProcess
   
   field :description, type: String
   field :sibling_level, type: Integer
-  field :duration, type: Integer
-  field :pause_duration, type: Integer
+  field :duration, type: Float
+  field :pause_duration, type: Float
   
   attr_accessible :sibling_level, :duration, :pause_duration, :description
